+++ src/api/video.ts
@@ -1,8 +1,9 @@
+import { RevError } from '../rev-error';
 import type { RevClient } from '../rev-client';
 import { Video, Rev } from '../types';
-import { isPlainObject } from '../utils';
 import { SearchRequest } from '../utils/request-utils';
-import { VideoReportRequest } from './video-report-request';
+import { videoReportAPI } from './video-report-request';
+import { videoDownloadAPI } from './video-download';
 
 export default function videoAPIFactory(rev: RevClient) {
     const videoAPI = {
@@ -25,8 +26,42 @@ export default function videoAPIFactory(rev: RevClient) {
         async details(videoId: string): Promise<Video.Details> {
             return rev.get(`/api/v2/videos/${videoId}/details`);
         },
+        /** get list of comments on a video */
+        async comments(videoId: string): Promise<Video.Comment[]> {
+            const response = await rev.get<Video.CommentListResponse>(`/api/v2/videos/${videoId}/comments`);
+            return response.comments;
+        },
+        async chapters(videoId: string): Promise<Video.Chapter[]> {
+            try {
+                const {chapters} = await rev.get<{chapters: Video.Chapter[]}>(`/api/v2/videos/${videoId}/comments`);
+                return chapters;
+            } catch (err) {
+                // if no chapters then this api returns a 400 response
+                if (err instanceof RevError && err.code === "NoVideoChapters") {
+                    return [];
+                }
+                throw err;
+            }
+        },
+        async supplementalFiles(videoId: string): Promise<Video.SupplementalFile[]> {
+            const {supplementalFiles} = await rev.get(`/api/v2/videos/${videoId}/supplemental-files`);
+            return supplementalFiles;
+        },
+        // async deleteSupplementalFiles(videoId: string, fileId: string | string[]): Promise<void> {
+        //     const fileIds = Array.isArray(fileId)
+        //         ? fileId.join(',')
+        //         : fileId
+        //     await rev.delete(`/api/v2/videos/${videoId}/supplemental-files`, { fileIds });
+        // },
+        async transcriptions(videoId: string): Promise<Video.Transcription[]> {
+            const {transcriptionFiles} = await rev.get(`/api/v2/videos/${videoId}/transcription-files`);
+            return transcriptionFiles;
+        },
         get upload() {
             return rev.upload.video;
+        },
+        async migrate(videoId: string, options: Video.MigrateRequest) {
+            await rev.put(`/api/v2/videos/${videoId}/migration`, options);
         },
         /**
          * search for videos, return as one big list. leave blank to get all videos in the account
@@ -73,39 +108,8 @@ export default function videoAPIFactory(rev: RevClient) {
             const { video } = await rev.get(`/api/v2/videos/${videoId}/playback-url`);
             return video;
         },
-        async download(videoId: string) {
-            const response = await rev.request<ReadableStream>('GET', `/api/v2/videos/${videoId}/download`, undefined, {
-                responseType: 'stream'
-            });
-            return response;
-        },
-        async downloadTranscription(videoId: string, language: string) {
-            const { body } = await rev.request<Blob>('GET', `/api/v2/videos/${videoId}/transcription-files/${language}`, undefined, { responseType: 'blob' });
-            return body;
-        },
-        async downloadThumbnail(query: string | { videoId?: string, imageId?: string; }) {
-            const {
-                videoId = '',
-                imageId = ''
-            } = typeof query === 'string'
-                    ? { imageId: query }
-                    : query;
-
-            if (!(videoId || imageId)) {
-                throw new TypeError('No video/image specified to download');
-            }
-
-            let thumbnailUrl = imageId
-                ? `/api/v2/media/videos/thumbnails/${imageId}.jpg`
-                // allow getting from api if only know the video ID
-                : (await videoAPI.playbackInfo(videoId)).thumbnailUrl;
-
-            const { body } = await rev.request<Blob>('GET', thumbnailUrl, undefined, { responseType: 'blob' });
-            return body;
-        },
-        report(options: Video.VideoReportOptions = {}) {
-            return new VideoReportRequest(rev, options);
-        }
+        ...videoDownloadAPI(rev),
+        ...videoReportAPI(rev)
     };
     return videoAPI;
 }